backend/pyproject.toml
[tool.uv]
package = true

[tool.setuptools]
py-modules = ["app"]

[project]
name = "periscope-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.15",
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "beautifulsoup4>=4.13.5",
    "brevo-python>=1.2.0",
    "email-validator>=2.3.0",
    "fastapi[standard]>=0.116.2",
    "feedparser>=6.0.12",
    "loguru>=0.7.3",
    "lxml>=6.0.1",
    "bcrypt>=4.2.0",
    "psycopg[binary]>=3.2.10",
    "pydantic-ai>=1.0.8",
    "pydantic-extra-types>=2.10.6",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "python-ulid>=3.1.0",
    "redis>=6.4.0",
    "sqlalchemy[asyncio]>=2.0.43",
    "sqlmodel>=0.0.25",
    "temporalio>=1.17.0",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.18.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.13.0",
    "pytest-asyncio>=1.2.0",
    "pytest-mock-resources>=2.12.4",
    "pytest-xdist>=3.8.0",
    "python-on-whales>=0.78.0",
    "fakeredis>=2.31.3",
    "pytest-timeout>=2.4.0",
]
